Add files via upload
This code was used to approach an activity of undergrad physics lab, related to fluid mechanics. Involving the behaviour of the tangential velocity of a vortex of a liquid containing water and glycerine in equal parts. At first it approaches the problem by adjusting the collected data by least squares, consequently it is compared with approaching the problem by the usage of a Physiscs Informed Neural Network obtaining a similar result for the viscosity nu. Finally there is an attempt to take into consideration the boundaries of the system in the PINN.